Diagrammatic representation of the problem.
Schematically, then, if time is a straight line, we have at some point date A and, at a later point, date B. Therefore, we want to calculate the "yellow" distance, the straight-line segment AB. What we will try to do is calculate AB indirectly, as the sum/difference of other segments, which are easier to compute. For example, the "blue" distance is very easy to calculate as it represents how many days have passed from New Year's Day of the year of date A to New Year's Day of the year in which date B falls. This is calculated as year B minus year A and multiplied by 365, since we want the result in days.

Similarly, we can relatively easily calculate the "green" and "orange" distances, that is, how many days have passed from the New Year's Day of date A to A and how many days have passed from the New Year's Day of date B to B. For example, if date A is 15/04 (any year), 15 days of April have passed since the New Year, along with the days of the 4-1 = 3 previous months.